requirements.txt
fastapi
uvicorn
PyMuPDF
pyserial
requests
reportlab
python-dotenv
pytest
